Logging configured and testing started
- Created a logging configuration file and added logging to <backup.py>, <list.py>, <restore.py>, and <cleaner.py>; logs are now written both to a log file and to the console
- Added more error checks, raised different exceptions, and logged these errors to the log file
- Started writing pytest tests for core logic files (currently only for <backup.py>), with approximately 87% test coverage:
     - tested all code using cProfile and pylint

+++ core/restore.py
@@ -1,12 +1,17 @@
 import json
 import shutil
 import os
+import logging
+
+logger = logging.getLogger(__name__)
 
 class RestoreFile:
     def __init__(self):
         self.data_file = os.path.join("database", "data.json")
 
     def restore_file(self, asked_id):
+        logger.info("Restoring file started")
+        logger.info("Restore requested for ID=%s", asked_id)
         with open(self.data_file, "r", encoding="utf-8") as f:
             data = json.load(f)
 
@@ -18,16 +23,21 @@ class RestoreFile:
                 continue
 
         if not found_block:
+            logger.info("BAD restore requested for ID=%s - check if file exists", asked_id)
             raise ValueError(f"No such file in the backup storage")
 
         original_path = found_block.get("Original path")
         backup_path = found_block.get("Backup path")
+        if not original_path or not backup_path:
+            logger.error("Invalid backup entry: %s", found_block)
+            raise ValueError("Backup entry is corrupted")
         try:
             shutil.copy2(backup_path, original_path)
             if os.path.exists(backup_path):
                 os.remove(backup_path)
-        except Exception as e:
-            print(f"Error during recovery {e}")
+        except Exception:
+            logger.exception("Error during recovery file with ID=%s", asked_id)
+            raise
 
         try:
             new_data = []
@@ -39,4 +49,7 @@ class RestoreFile:
             with open(self.data_file, "w", encoding="utf-8") as f:
                 json.dump(new_data, f, ensure_ascii=False, indent=4)
         except Exception as e:
-            print(f"Error during database changes {e}")
+            logger.exception(f"Error during database changes {e}")
+            raise
+
+        logger.info("Finished restoring file")
